requirements.txt
absl-py>=1.4.0
anyio>=3.6.2
argon2-cffi>=21.3.0
argon2-cffi-bindings>=21.2.0
arrow>=1.2.3
ase>=3.22.1
asttokens>=2.2.1
attrs>=22.2.0
backcall>=0.2.0
beautifulsoup4>=4.11.2
bleach>=6.0.0
cachetools>=5.3.0
cffi>=1.15.1
chainer>=7.8.1
chainer-chemistry>=0.7.1
charset-normalizer>=3.0.1
colorama>=0.4.6
comm>=0.1.2
contourpy>=1.0.7
cycler>=0.11.0
debugpy>=1.6.6
decorator>=5.1.1
deepchem>=2.7.1
defusedxml>=0.7.1
descriptastorus>=2.6.0
deepchem
easydict>=1.10
EFGs>=0.8.4
executing>=1.2.0
fastjsonschema>=2.16.2
filelock>=3.9.0
fonttools>=4.38.0
fqdn>=1.5.1
google-auth>=2.17.3
google-auth-oauthlib>=1.0.0
graphviz>=0.20.1
grpcio>=1.54.0
icecream>=2.1.3
idna>=3.4
imageio>=2.25.0
importlib-metadata>=6.0.0
importlib-resources>=5.10.2
ipykernel>=6.21.1
ipython>=8.9.0
ipython-genutils>=0.2.0
ipywidgets>=8.0.4
isoduration>=20.11.0
jedi>=0.18.2
Jinja2>=3.1.2
joblib>=1.2.0
jsonpointer>=2.3
jsonschema>=4.17.3
jupyter>=1.0.0
jupyter-console>=6.4.4
jupyter-events>=0.6.3
jupyter_client>=8.0.2
jupyter_core>=5.2.0
jupyter_server>=2.2.1
jupyter_server_terminals>=0.4.4
jupyterlab-pygments>=0.2.2
jupyterlab-widgets>=3.0.5
kiwisolver>=1.4.4
lazy_loader>=0.1
libauc>=1.2.0
littleutils>=0.2.2
Markdown>=3.4.3
MarkupSafe>=2.1.2
matplotlib>=3.6.3
matplotlib-inline>=0.1.6
mistune>=2.0.4
MolVS>=0.1.1
mpmath>=1.3.0
nbclassic>=0.5.1
nbclient>=0.7.2
nbconvert>=7.2.9
nbformat>=5.7.3
nest-asyncio>=1.5.6
networkx>=3.1
notebook>=6.5.2
notebook_shim>=0.2.2
numpy>=1.24.1
oauthlib>=3.2.2
ogb>=1.3.6
opencv-python>=4.7.0.68
outdated>=0.2.2
packaging>=23.0
pandas>=1.5.3
pandas-flavor>=0.5.0
pandocfilters>=1.5.0
parso>=0.8.3
pexpect>=4.8.0
pickleshare>=0.7.5
Pillow>=9.4.0
pkgutil_resolve_name>=1.3.10
platformdirs>=3.0.0
POT>=0.9.0
prometheus-client>=0.16.0
prompt-toolkit>=3.0.36
protobuf>=3.20.1
psutil>=5.9.4
ptyprocess>=0.7.0
pure-eval>=0.2.2
pyaml>=23.5.7
pyasn1>=0.5.0
pyasn1-modules>=0.3.0
pycparser>=2.21
Pygments>=2.14.0
pyparsing>=3.0.9
pyrsistent>=0.19.3
python-dateutil>=2.8.2
python-json-logger>=2.0.4
pytz>=2022.7.1
PyWavelets>=1.4.1
PyYAML>=6.0
pyzmq>=25.0.0
qtconsole>=5.4.0
QtPy>=2.3.0
rdkit>=2022.9.4
requests>=2.28.2
requests-oauthlib>=1.3.1
rfc3339-validator>=0.1.4
rfc3986-validator>=0.1.1
rsa>=4.9
scikit-image>=0.19.3
scikit-learn>=1.2.0
scipy==1.5.4
seaborn>=0.12.2
Send2Trash>=1.8.0
six>=1.16.0
scikit-learn
sniffio>=1.3.0
soupsieve>=2.3.2.post1
stack-data>=0.6.2
sympy>=1.11.1
tensorboard>=2.13.0
tensorboard-data-server>=0.7.0
tensorboardX>=2.5.1
terminado>=0.17.1
texttable>=1.6.7
threadpoolctl>=3.1.0
tifffile>=2023.2.3
tinycss2>=1.2.1
torchviz>=0.0.2
tornado>=6.2
tqdm>=4.64.1
traitlets>=5.9.0
typing>=3.7.4.3
typing_extensions>=4.4.0
uri-template>=1.2.0
urllib3>=1.26.14
wcwidth>=0.2.6
webcolors>=1.12
webencodings>=0.5.1
websocket-client>=1.5.1
Werkzeug>=2.3.3
widgetsnbextension>=4.0.4
xarray>=2023.1.0
zipp>=3.12.1
